requirements.txt
langchain==1.2.10
langchain-community==0.4.1
fastapi==0.135.0
uvicorn==0.41.0
chromadb==1.5.2
pypdf==6.7.4
python-dotenv==1.2.2
tiktoken==0.12.0
google-generativeai==0.8.6
python-multipart==0.0.22
langchain-google-genai==4.2.1
python-dotenv==1.2.2
pydantic-settings==2.13.1
aiofiles==25.1.0
PyMupdf==1.27.1

